generalize distribution away from socrata-specific
